Add files via upload
the script does have section of file navigation specific to the domain i'm on so this would need to be modified if used anywhere else. it pulls the active logged in username with the domain leading before removing the exact domain name. this is done in the clear_cache function at the start