Update browser_cleaner.ps1
updated script to no longer pull currently logged in user but instead to set an environmental variable to the localappdata and tack that to the front of the needed folder paths. doing it this way means while running locally or remotely such as through Tanium this should work in both cases.